Move tests for problem 1.2 to combined chapter test.
